Don't log "affected rows" when exactly 1 row is affected.
The "affected rows" on each Post _is_ useful, but it takes a lot of space.
Would be nice if it could fit on the same line as the query itself.
Perhaps a "execution took xxx msecs" could be added on the same line for each query too.

For now, it seems odd that there's a fine row counter for dbgrid Post queries, but not for everything else, doesn't it?
Isn't that just really inconsistent?
I've removed the log line on these grounds..

If you had plans with it, or just think it really should be there, add it back..