Patch by: fco_ernesto@yahoo.com.br
Feature: <Ctrl>+<Up> and <Ctrl>+<Down> moves entries in the 'tables to export' list in the export dialog up and down.

Users can now customize the order in which tables are exported, which is useful because it allows tables with simple foreign key relations to be exported in an order that allows them to be correctly imported again.