Fix Zeos-internal queries to use ZConnection.Catalog-property only if not empty. The datagrid is now working again.
The everlasting problem are this Catalog-property, for which we had a fix in the old Zeos-version and now we have a new fix in the newer Zeos-version. This new fix is even worse than the old one because it affects roughly 60 new lines, where the old fix was only a handful lines big.

Obviously HeidiSQL's datagrid works this way, but a *MUCH* better way would be to find out where we have to simply set the catalog-property and revert this change. I tried to figure it out, but I'm failing in understanding how the MysqlQueryThread uses a ZConnection and its properties. Joey??

The only thing I definitely know for now is that Zeos depends on the Catalog-property of a ZConnection, where I always thought it depends on the Database-property, which is in my eyes quite the same, only a different name for the same thing on different DBMSs.