Fix issue #1685950: "copy options available when no data in grid" by centralising the enabled/disabled-commands in one procedure.
Positive side-effect: "Delete record(s)" and "View as HTML" menuitems are also disabled on empty grids

Negative: procedure-parameter FrmIsFocussed is somehow unclear in its meaning.

Leave open for now:
"Scrolling with the mouse wheel in the query grid yields the same error message" (ZQuery1: Cannot perform this operation on a closed dataset.).
This is a second issue and has to be fixed by unlinking the ZQuery1 to Datasource2 in case of an SQL-error.