Turn the query progress window not to be "always on top". It is already modal as set by the callers so that ensures no internal windows can cover this window.
I see no obvious reason why this status window should cover windows of other applications. And it can be annoying if you have a slow query running and want to work with other applications in the meantime. I asked the original developer for the initial reason for that change, but got no reply yet. Likely I won't get an answer ever.