fix: flashing main form in background after cancel session manager
Decided to fully show the main form before showing the session manager through an Application.OnIdle event. Hoping no user manages to click somewhere on the main form, causing crashes. The only reasonable alternative was to show the session manager before TMainForm is created, but this would be more effort.